Code/awa2/requirements.txt
torch
torchvision
pandas
scikit-learn
Pillow
tqdm
